+++ R/dataU.R
@@ -1,0 +1,17 @@
+#' Artificial, small scanner data set  
+#'
+#' A collection of artificial scanner data on 6 products sold in Dec, 2018. Product descriptions contain the information about their grammage and unit.
+#' @format A data frame with 5 columns and 6 rows. The used variables are as follows:
+#'
+#' \code{time} - Dates of transactions (Year-Month-Day)
+#'
+#' \code{prices} - Prices of sold products [PLN]
+#'
+#' \code{quantities} - Quantities of sold products [item]
+#'
+#' \code{prodID} - Unique product codes 
+#'
+#' \code{description} Descriptions of sold products (data set contains 6 different product descriptions)
+#' @docType data
+#'
+"dataU"