+++ NEWS.md
@@ -1,3 +1,18 @@
-# PriceIndices 0.0.0.9000
+# Package update 0.0.1 -> 0.0.2
 
-* Added a `NEWS.md` file to track changes to the package.
+
+## Improvements
+
+* Function **data_matching()** has been modified and now it is much more faster,
+
+* Examples concerning **chain indices** have been modified and now they are less time-consuming while checking,
+
+* New data set is added (**dataU**)
+
+* New functions for scanner data processing are added (**data_unit** and **data_norm**)
+
+## Bug fixes
+
+* Access to HASP and WISP methods has been added in the **price_index()** function and now **price_indices()** works correctly,
+
+* The default value and the meaning of the **sensitivity** parameter in **data_preparing()** and **model_classification()** functions have been corrected.